64-bit crash fix

+++ gtkwave3/src/ae2.c
@@ -194,7 +194,7 @@ static unsigned long ae2_read_symbol_rows_2(AE2_HANDLE handle, unsigned long sym
 {
 if(symbol_idx <= GLOBALS->ae2_num_facs)
 	{
-	unsigned long r = ae2_read_symbol_rows(handle, symbol_idx);
+	unsigned long r = (unsigned long)((unsigned int)ae2_read_symbol_rows(handle, symbol_idx));
 	if(r > AE2_MAX_ROWS) r = AE2_MAX_ROWS;
 	return(r);
 	}
